misc: add facilities to insert keys to e4crypt
This patch adds the following options to the e4crypt:
-a keyring [-k keyref] [-p passphrase] [-b hexstring]
-r keyring -k keyref

Where '-a' allows to add a key to a 'keyring', where key is derived
from user typed passphrase (if both '-p' and '-b' are omitted), or
from 'passphrase' specified by '-p' option, or stored 'as-is' if
option '-b' is used.

Command '-r' provides means for deleting a key from a keyring.

Signed-off-by: Ildar Muslukhov <muslukhovi@gmail.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>